Only use empty chains for overlapped reads.
This avoids the need to perform a deep copy of the last chain with
data in evbuffer_add_buffer if the chain is pinned for reading. The
patch adds a bool argument, 'empty_only', to _evbuffer_expand_fast
and _evbuffer_read_setup_vecs.

The current behavior of evbuffer_add_buffer is to move pinned chains
from the source buffer to the destination buffer, which is incorrect.
The next patch will attempt to fix this.